Update GitPrRequestGeneration function in gh.vim
The GitPrRequestGeneration function in gh.vim now pipes the git diff
command output to cat.
In addition, the prompt message has been updated to have a revised
format for the pull request.
Specifically, the change calls for a title for the pull request instead
of a descriptive title in the pull request message.