Commit: Update docker build and modify file structure handling
Relocated the Docker login command to execute before building Docker
images. The Docker login command should be executed prior to the Docker
build command to ensure that the login is successful before the building
process starts.

Replaced "vault" with "mount_files" to improve clarity in the file
structure. Previously, "vault" was the named location for mounted files,
this has been changed to "mount_files", which better indicates its
purpose.

Added functionality to the script allowing for setup, build and
configuration tasks. The new commands (setup, build, config) aid the
user in managing the mounting structure, building the vim container, and
editing the runit files, respectively. This enables a cleaner and more
efficient workflow.