Update Dockerfile and improve compatibility for Vim configs
Updated the Dockerfile to use ubuntu:22.04 and to install necessary
packages in combined commands.

Also added a daily vim PPA for more frequent updates.

In the run_vim function, added steps to pass the Display server to the
docker environment for better compatibility with GUI-based applications.

Modified command in the run_build function to navigate to script
location before executing it.

Updated vim configurations to improve compatibility and usability.

Refactored 'nocompatible' command in vim configs for proper sourcing.

Fixed indentation in AI.vim file and made additions to clipboard and
backspace settings in vim.vim file for better user experience.