Added pyperclip to Dockerfile and adjusted run_config
This commit includes the addition of the 'pyperclip' Python package to
the Dockerfile.

The 'run_config' command in the 'runit' file has now been slightly
adjusted for better command-handling.

Furthermore, a Docker container pull command and local xhost command for
root were added to 'setup' file.

This improves setup automation and enables clipboard access for Python
in Docker.