Added methods for secret and git gist functionality
This commit adds new methods to the Gist and Secret classes, providing
additional functionality.
Within the 'Secret' class, methods were added to construct an instance
of the class from a gist, delete secret descriptions, copy the secret
path onto clipboard and open the file/directories via integrated
command.
For the 'Gist' class, methods were added to create a Gist instance from
a Secret instance, copy gist link path and open the gist URL in a
browser.
Additionally, the commit includes the removal of unnecessary numpy and
scipy import statements and adds the import for pyperclip which enables
copying to the clipboard.
It also improves the action for deletion and transitioning from public
to private.
The visual representation of descriptions in the '__repr__' function was
modified to provide clearer output.