Added argparse to handle path/reader flag
This commit introduces the argparse module to better handle command line
arguments for the open.py script.

It replaces the manual sys.argv method previously used for obtaining the
path to be processed.

The script now also accepts a --reader flag to modify URLs.

Instead of directly executing the path-to-command functionality in the
main part of the script, we now execute a main function that entails
handling of the argparser and the path to command operation.