Summary: Refactored gistnotes and adjusted runit scripts.
In this commit, several revisions have been made to the `gistnotes.py`
file.
The existing code regarding the `Vault` and `Secret` classes was removed
and two new classes, `Wiki` and `Note`, were added.
The focus of the changes centers around retrieving a description from a
markdown file.
The `runit` scripts were also adjusted to implement these new changes.
There were some minor changes in the `note_default.py` and `open.py`
scripts.
The overall project seems to be moving from a focus on Gists and Secrets
towards a wiki with notes.