Update: Added calendar functionality and modified scripts
Added a calendar functionality to Vimwiki using a new plugin, along with
modifications to the scripts 'gistnotes.py' and 'runit'.
Changed the keyboard shortcuts in coc.vim and made adjustments to handle
the addition of calendar in windows.vim.
Changes in 'runit' include modification in the run_gist_notes and
run_local functions, where the command value was set to
'/scripts/gistnotes.py; vim;' and 'vim;' respectively.
In 'gistnotes.py', a return statement was added before executing print
statement.
The function 'LeftBarToggle' was also updated in windows.vim to consider
'calendar' Filetype.
Finally, a new function 'LeftBarToCalendar' was added to wiki.vim to
toggle is between usual view and the calendar.