"Updated key mapping and functions in vim configs"
Updated key mappings in coc.vim and wiki.vim.
Modified function MoveRight() in windows.vim.
Removed mapping in wiki.vim and a condition in windows.vim.

In coc.vim, remapped the key for 'coc-definition' shortcut from <cr> to
gd.

In wiki.vim, removed key mapping for 'VimwikiSplitLink' and updated the
key mapping for 'LeftBarToCalendar()' from 'C' to 'c'.
Added new key mapping 'F' for 'LeftBarToNerdFind()'.

For windows.vim, in the function MoveRight(), 'calendar' and 'qf' have
been added to the 'exempt_ft' and 'left_bar_ft' arrays.
Also, a specific condition for 'qf' filetype was removed.

Lastly, 'calendar' was added to the 'ft' conditions in the
LeftBarToggle() function.