Update print statements and vim auto commands
In the `note_default.py` file, the `print` statements inside two loops
were refactored for cleaner code.

Previously, an empty string was printed after every event and todo.

This has been changed to print only once after looping through all the
events and todos.

In `windows.vim`, the autocmd for `MoveRight()` has been modified.

Previously it was called whenever a buffer was read or new file was
made.

Now, it's invoked for any Filetype, increasing the versatility of the
function.