Updated autocomplete and wiki vim setups
Expanded the autocomplete.vim to include a register_source for
'vimwiki'.
This includes the 'name', 'allowlist', and 'completor' fields required
to set up the source.
Also, made a few changes in the wiki.vim configuration.
Now the auto_tags field in the wiki_1 dictionary is enabled for
auto-tagging.
Removed the line that globally applied a vimwiki extension.
This results in a more customized autocomplete functionality for vimwiki
and a fine-tuned wiki set-up.