Add R language server to lsp.vim
A R language server has been added to the lsp.vim configuration file.
A run script for the R language server 'r-languageserver' has also been
created.
This allows for the execution of R in a slave mode, using the
'languageserver::run()' function.
The server is registered and allowlisted for 'r' in the lsp_setup.