Updated Dockerfile, autocomplete.vim, and lsp.vim
This commit makes updates to the Dockerfile, autocomplete.vim, and
lsp.vim files.
The Dockerfile's changes include tweaking the list of installed packages
and adding new ones, such as ipython and ipykernel in the python
packages installation.
There have also been additions of the lines necessary for the
installation of
Ltex-ls and Quarto.
In the autocomplete.vim file, a new line setting
`g:asyncomplete_popup_delay` to  500 has been inserted.
Lastly, the lsp.vim file has been updated with new configurations,
including the addition of log files, the replacement of 'r' with
'r-languageserver', and the registration of the Ltex-ls server.