Update terminal automation in Vim
Made adjustments to the terminal automation behavior in Vim.
Previously, on terminal close, 'close!' was executed automatically.
This has been changed to only delete the buffer if the status event is not active.
A startinsert command got added to execute on BufWinEnter and WinEnter events.
In the StartTerm function, the command to execute in the terminal is now passed as an argument.
Terminal commands were changed to support the new StartTerm signature.
A new keybinding got added for running commands with parameters.