Set tmux-256color in Dockerfile, and update filetype in wiki.vim
The Dockerfile was updated to set the TERM variable to "tmux-256color".
This allows for correct interpretation of color and styling codes by tmux sessions.
White space was also added at the end of lsp.lua, but this won't affect any functionality.

In the wiki.vim file, the auto command associated with FileType 'vimwiki' was changed.
Instead of setting the local syntax to 'quarto', it now sets the filetype to 'quarto'.
This subtle change aligns its behavior better with conventionally expected vim behaviors.