Fixed isinstance argument order in todays_events
In todays_events method of Calendar class, the order of arguments in the isinstance function was incorrectly put.
The original line was checking if the 'date' is an instance of 'day', which doesn't make sense.
This change corrects the order of arguments to correctly check if 'day' is an instance of 'date'.